More refactoring, a bit more focus on consolidating different header files and structures
